Refactored some parallel code and improved coverage reports.
